app/agent/pyproject.toml
[project]
name = "secundus-stylist-agent"
version = "0.1.0"
description = "Standalone stylist agent — Gemini ReAct loop calling the Secundus backend over HTTP"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-adk>=0.4.0",
    "python-socketio[client]>=5.11.0",
]

[tool.uv]
package = false
